+++ CHANGELOG.md
@@ -3,6 +3,11 @@ CHANGELOG
 
 * 0.4.0 (xx-xx-xxxx)
 
+* 0.3.3 (xx-xx-2013)
+
+  * Add convenient Stream::getDimensions method to extract video dimension.
+  * Add DisplayRatioFixer Frame filter.
+
 * 0.3.2 (08-08-2013)
 
   * Fix A/V synchronization over flash and HTML5 players.